@@ -48,19 +48,19 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@angular/animations": "^9.1.6",
+    "@angular/animations": "^10.2.0",
     "@angular/cdk": "^9.2.3",
-    "@angular/common": "^9.1.6",
-    "@angular/compiler": "^9.1.6",
-    "@angular/core": "^9.1.6",
+    "@angular/common": "^10.2.0",
+    "@angular/compiler": "^10.2.0",
+    "@angular/core": "^10.2.0",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "^9.1.6",
+    "@angular/forms": "^10.2.0",
     "@angular/material": "^9.2.3",
     "@angular/material-moment-adapter": "^9.2.3",
-    "@angular/platform-browser": "^9.1.6",
-    "@angular/platform-browser-dynamic": "^9.1.6",
-    "@angular/platform-server": "^9.1.6",
-    "@angular/router": "^9.1.6",
+    "@angular/platform-browser": "^10.2.0",
+    "@angular/platform-browser-dynamic": "^10.2.0",
+    "@angular/platform-server": "^10.2.0",
+    "@angular/router": "^10.2.0",
     "@cfstratos/ajsf-material": "^0.1.6",
     "@cfstratos/monaco-yaml": "^2.5.0",
     "@ngrx/effects": "^9.1.2",
@@ -102,13 +102,13 @@
   "devDependencies": {
     "@angular-builders/custom-webpack": "^9.1.0",
     "@angular-devkit/architect": "^0.901.7",
-    "@angular-devkit/build-angular": "^0.901.12",
-    "@angular-devkit/build-ng-packagr": "~0.901.5",
-    "@angular-devkit/core": "^9.1.7",
-    "@angular-devkit/schematics": "^9.1.5",
-    "@angular/cli": "^9.1.5",
-    "@angular/compiler-cli": "^9.1.6",
-    "@angular/language-service": "^9.1.6",
+    "@angular-devkit/build-angular": "^0.1002.0",
+    "@angular-devkit/build-ng-packagr": "~0.1002.0",
+    "@angular-devkit/core": "^10.2.0",
+    "@angular-devkit/schematics": "^10.2.0",
+    "@angular/cli": "^10.2.0",
+    "@angular/compiler-cli": "^10.2.0",
+    "@angular/language-service": "^10.2.0",
     "@schematics/angular": "^9.1.5",
     "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "~2.0.8",
@@ -143,7 +143,7 @@
     "lodash": "^4.17.19",
     "mem": "6.1.0",
     "mktemp": "^1.0.0",
-    "ng-packagr": "^9.1.1",
+    "ng-packagr": "^10.1.2",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
     "protractor": "^5.4.3",
@@ -155,7 +155,7 @@
     "sass-lint": "^1.12.1",
     "stratos-protractor-reporter": "^1.2.3",
     "ts-node": "^8.8.2",
-    "tslint": "~6.1.1",
-    "typescript": "3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "4.0.5"
   }
 }
