@@ -134,7 +134,7 @@ export class ExtensionService {
 
   // Declare extensions - this is a trick to ensure the Angular Build Optimiser does not
   // optimize out any extension components
-  public static declare(components: any[]): ModuleWithProviders {
+  public static declare(components: any[]): ModuleWithProviders<ExtEmptyModule> {
     return {
       ngModule: ExtEmptyModule
     };
